+++ mcp/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 # [0.2.1] 2025-05-07
 
 - Add `readonly` option to prevent submitting orders to the exchange.
+- Add support market temperature api.
 
 # [0.2.0] 2025-04-23